requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
django-datatables-view==1.19.1
django-enumfields==2.0.0
django-filter==2.3.0
djangorestframework==3.11.0
Markdown==3.2.2
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
